more changes in response to code review

+++ test/intest1.F90
@@ -25,7 +25,7 @@ program intest1
        bfmg(mxbf), cds3(mxds3)*6, tagpr*8, celem*60, cunit*22
   character*20 filnam / 'testfiles/IN_1' /
   character filost / 'r' /
-  integer i4dy, idxmsg
+  integer*4 i4dy, idxmsg
 
   equivalence (bfmg(1), ibfmg(1))
